# perf: reduce B/op and allocs/op