benchstat:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/decoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                 │   old.txt    │              new2.txt               │
                 │    sec/op    │   sec/op     vs base                │
CheckIntegrity-4   18.59m ± 11%   16.57m ± 1%  -10.82% (p=0.007 n=10)

                 │  old.txt   │              new2.txt              │
                 │    B/op    │    B/op     vs base                │
CheckIntegrity-4   314.0 ± 4%   149.0 ± 6%  -52.55% (p=0.000 n=10)

                 │  old.txt   │              new2.txt              │
                 │ allocs/op  │ allocs/op   vs base                │
CheckIntegrity-4   9.000 ± 0%   1.000 ± 0%  -88.89% (p=0.000 n=10)

```